endpointv2

+++ src/data/FollowersDatabase.ts
@@ -4,13 +4,24 @@ export class FollowersDatabase extends BaseDB {
   private static TABLE_NAME = "Followers";
 
   public async followerUser(idUser: string, idFollower: string): Promise<void> {
-    await this.getConnection()
-      .insert({
-        idUser,
-        idFollower
-      })
+    try{
+      const result = await this.getConnection()
+      .select("*")
       .into(FollowersDatabase.TABLE_NAME);
 
-    await BaseDB.destroyConnection()
+      if(result[0] !== undefined) {
+        throw new Error ("Já segue");
+      }
+        await this.getConnection()
+          .insert({
+            idUser,
+            idFollower
+          })
+          .into(FollowersDatabase.TABLE_NAME);
+
+        await BaseDB.destroyConnection();
+      } catch (error) {
+        return error.message;
+      }
   }
 }
